Tpetra::CrsGraph::getNodeNumEntries: Fix CUDA build error
@trilinos/tpetra Here is an example build report:

https://testing.sandia.gov/cdash/viewBuildError.php?buildid=3041592

.../Trilinos/packages/kokkos/core/src/Kokkos_Serial.hpp(448): error: calling a __device__ function("operator()") from a __host__ function("exec") is not allowed
          detected during:
...
            instantiation of "void Kokkos::parallel_reduce(const std::string &, const PolicyType &, const FunctorType &, ReturnType &, Kokkos::Impl::enable_if<Kokkos::is_execution_policy<PolicyType>::value, void>::type *) [with PolicyType=Kokkos::RangePolicy<Kokkos::HostSpace::execution_space, int>, FunctorType=lambda [](const LO &, size_t &)->void, ReturnType=size_t]"
.../Trilinos/packages/tpetra/core/src/Tpetra_CrsGraph_def.hpp(906): here
...

What's happenening is that for CUDA < 8.0, KOKKOS_LAMBDA only marks
the lambda as __device__ (this is a CUDA issue, not a Kokkos issue).
However, the code in question is supposed to run on host.  This
explains the host code (Kokkos::parallel_reduce with host execution
space) attempting to call a __device__ function (the lambda, marked
with KOKKOS_LAMBDA).

I could have fixed this issue just by requiring CUDA >= 8.0.  In that
case, Kokkos defines KOKKOS_LAMBDA with both __host__ and __device__
markings.  However, the code in question will only ever run on host,
so I just fixed it by replacing KOKKOS_LAMBDA with the usual host
capture clause [=].

I tried changing Tpetra::CrsGraph::k_numRowEntries_ from a HostMirror
(currently, a UVM allocation) to an actual host View (of
Kokkos::HostSpace), but that caused many tests to fail in nonobvious
ways.  Thus, I left it as a HostMirror allocation for now, and added
a comment warning other developers about the issue.