Tpetra: Fix #1571
@trilinos/tpetra Tpetra's CMake logic now attempts to detect whether
the MPI implementation is CUDA aware.  If automatic detection does not
succeed, Tpetra just makes the safe assumption that MPI is not CUDA
aware.

Currently, automatic detection requires OpenMPI.  If not using
OpenMPI, Tpetra conservatively assumes lack of CUDA awareness.  It
would be wise for us to extend detection to support other MPI
implementations, but for now, this covers a common use case for
Trilinos testing.

Automatic detection depends on running an executable.  This is
relevant for cross compilation, so I have added two measures to
protect against misleading results in that case:

  1. If CMAKE_CROSSCOMPILING is ON, Tpetra skips detection and prints
     a configure-time message telling users that they may set
     Tpetra_ASSUME_CUDA_AWARE_MPI explicitly if they want Tpetra to
     assume that MPI is CUDA aware.

  2. If users set Tpetra_ASSUME_CUDA_AWARE_MPI explicitly, Tpetra
     skips detection and assumes the user's value as the default.

Thanks to @ibaned for the review and suggestions!  I incorporated
these into the original pull request #1602.